+++ macrostrat-cli/macrostrat_cli/io/criticalmaas/steps.py
@@ -1,102 +1,10 @@
-"""
-Write a Macrostrat database to a GeoPackage file using the GeoPandas library.
-
-Currently, this represents the Macrostrat internal PostGIS schema more or less unmodified.
-As a next step, we will rework it to fit the CriticalMAAS TA1 output schema.
-https://github.com/DARPA-CRITICALMAAS/schemas/tree/main/ta1
-"""
-
 from macrostrat.database import Database
 from macrostrat.database.mapper import BaseModel
-from pathlib import Path
-from rich import print
-from criticalmaas.ta1_geopackage import GeopackageDatabase
 from shapely.geometry import mapping
 from geoalchemy2.elements import WKBElement
 from geoalchemy2.shape import to_shape
-from pydantic import BaseModel
 from typing import Generator
-from psycopg2.sql import Identifier
-
-
-def write_map_geopackage(
-    db: Database, identifier: str | int, filename: str = None, overwrite: bool = False
-):
-    """Write a Macrostrat map dataset (stored in PostGIS) to a GeoPackage file using GeoPandas and SQLAlchemy."""
-
-    _map = get_map_identifier(db, identifier)
-
-    if filename is None:
-        filename = Path(f"{_map.slug}.gpkg")
-    _unlink_if_exists(filename, overwrite=overwrite)
-
-    print(f"Map [bold cyan]{_map.slug}[/] (#{_map.id}) -> {filename}")
-
-    # Create the GeoPackage
-    gpd = GeopackageDatabaseExt(filename)
-
-    # Get models automapped from the Geopackage schema
-    gpd.automap()
-    Map = gpd.model.map
-    MapMetadata = gpd.model.map_metadata
-    LineType = gpd.model.line_type
-    PolygonType = gpd.model.polygon_type
-    GeologicUnit = gpd.model.geologic_unit
-    PointType = gpd.model.point_type
-
-    ### MAP metadata ###
-
-    gpd.write_models(_build_map_metadata(db, _map, Map, MapMetadata))
-
-    ### LINE FEATURES ###
-
-    # Insert the line types
-    valid_types = gpd.enum_values("line_type")
-    gpd.write_models(_build_line_types(db, _map, LineType, valid_types))
-    gpd.write_features("line_feature", _build_line_features(db, _map))
-
-    ### POINT FEATURES ###
-    valid_types = gpd.enum_values("point_type")
-    gpd.write_models(_build_point_types(db, _map, PointType, valid_types))
-    gpd.write_features("point_feature", _build_point_features(db, _map))
-
-    ### POLYGON FEATURES ###
-
-    gpd.write_models(_build_polygon_types(db, _map, PolygonType, GeologicUnit))
-    gpd.write_features("polygon_feature", _build_polygon_features(db, _map))
-
-
-# Helper classes
-
-
-class GeopackageDatabaseExt(GeopackageDatabase):
-    def write_models(self, models: list):
-        self.session.add_all(models)
-        self.session.commit()
-
-    def enum_values(self, enum_name: str):
-        """Get the values of an enum type."""
-        table_name = "enum_" + enum_name
-        try:
-            model = getattr(self.model, table_name)
-        except AttributeError:
-            raise ValueError(f"Enum type {enum_name} does not exist")
-
-        query = self.session.query(model.name)
-        # Insert the line types
-        return set(query.all())
-
-
-class MapIdentifier(BaseModel):
-    """A Macrostrat map identifier."""
-
-    id: int
-    slug: str
-    url: str = None
-    name: str = None
-
-
-# HELPER METHODS FOR SPECIFIC STEPS
+from .helpers import MapIdentifier
 
 
 def _build_map_metadata(
@@ -308,33 +216,3 @@ def _build_polygon_features(db, _map: MapIdentifier):
         )
         geometry = mapping(to_shape(WKBElement(vals.pop("geom"))))
         yield {"geometry": geometry, "properties": vals}
-
-
-### HELPER METHODS ###
-
-
-def get_map_identifier(db, identifier: str | int) -> MapIdentifier:
-    """Get a map identifier from a map ID or slug."""
-    query = "SELECT source_id, slug, name, url FROM maps.sources"
-    params = {}
-    try:
-        map_id = int(identifier)
-        query += " WHERE id = %(source_id)s"
-        params["source_id"] = map_id
-    except ValueError:
-        map_slug = identifier
-        query += " WHERE slug = %(slug)s"
-        params["slug"] = map_slug
-
-    res = next(db.run_sql(query, params=params)).first()
-
-    return MapIdentifier(id=res.source_id, slug=res.slug, url=res.url, name=res.name)
-
-
-def _unlink_if_exists(filename: Path, overwrite: bool = False):
-    file_exists = filename.exists()
-    if file_exists and not overwrite:
-        raise FileExistsError(f"File {filename} already exists")
-
-    if file_exists and overwrite:
-        filename.unlink()
